Fix compiler warning with bit shifts.
The ihr_data field points to uint8_t data, which the compiler
complains about when shifting by a large bit count.  Add explicit cast
to uint32_t for correctness.

+++ src/ihex_record.c
@@ -108,7 +108,8 @@ int ihex_rs_iterate_data(ihex_recordset_t* rs, uint_t *i, ihex_record_t **rec, u
 				
 				break;
 			case IHEX_ELA:
-				offset = (x->ihr_data[0] << 24) + (x->ihr_data[1] << 16);
+				offset = ((uint32_t) x->ihr_data[0] << 24)
+					+ ((uint32_t) x->ihr_data[1] << 16);
 				if (off)
 				{
 					*off = offset;
